# PortalMobileApp/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ITC Student Portal API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/departments": {
            "get": {
                "tags": [
                    "Departments"
                ],
                "summary": "List all departments",
                "operationId": "da8820891dd6bf08713e9d77fbe6389f",
                "responses": {
                    "200": {
                        "description": "List of departments"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Departments"
                ],
                "summary": "Create a new department",
                "operationId": "c0f48115550ccf6a76b43ff263bfc05c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Information Technology"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "IT"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "IT Department"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Department created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/departments/{id}": {
            "put": {
                "tags": [
                    "Departments"
                ],
                "summary": "Update department",
                "operationId": "1d6fe343b0c809ce2a0722693aed793d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Computer Science"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "CS"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "CS Department"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Department updated successfully"
                    },
                    "404": {
                        "description": "Department not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Departments"
                ],
                "summary": "Delete department",
                "operationId": "5cc7aebc708fbbbeb04b81504f62b499",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Department deleted successfully"
                    },
                    "404": {
                        "description": "Department not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/timetable/user/{user_id}": {
            "get": {
                "tags": [
                    "Timetable"
                ],
                "summary": "Get timetable for a student (by user)",
                "operationId": "f862c043736de738855ddd3175fea9d0",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of timetable for student's group"
                    }
                }
            }
        },
        "/api/timetable/group/{group_id}": {
            "get": {
                "tags": [
                    "Timetable"
                ],
                "summary": "Get timetable by group",
                "operationId": "c73f093e025d11b04b242be42bbef453",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of timetable for a group"
                    }
                }
            }
        },
        "/api/timetable": {
            "post": {
                "tags": [
                    "Timetable"
                ],
                "summary": "Create a timetable",
                "operationId": "a124470b3cd2090c1b4da8df4b9a0acd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "group_id",
                                    "class_id",
                                    "day_of_week",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "group_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "class_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "semester_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "subject_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "day_of_week": {
                                        "type": "string",
                                        "example": "Mon"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Network Fundamentals"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "08:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "10:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Timetable created successfully"
                    }
                }
            }
        },
        "/api/timetable/{id}": {
            "put": {
                "tags": [
                    "Timetable"
                ],
                "summary": "Update a timetable",
                "operationId": "27211480240a138d1a2311265266b095",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Title"
                                    },
                                    "day_of_week": {
                                        "type": "string",
                                        "example": "Tue"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "09:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "11:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Timetable updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Timetable"
                ],
                "summary": "Delete a timetable",
                "operationId": "d4251b115a8c370d6f59941e0c3aea99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Timetable deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}